fix: consolidate catalog lookups in the core and right-size the drift pins
